backend/requirements.txt
fastapi==0.118.3
uvicorn==0.37.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.16.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
groq==0.11.0
pinecone-client==3.2.2
langchain==0.3.27
sentence-transformers==5.1.1
python-dotenv==1.1.1
email-validator==2.1.1
pydantic[email]==2.12.0
pydantic-settings==2.7.1
